+++ src/components/add-ticket/forms/FormNeedTransport.tsx
@@ -17,6 +17,8 @@ import { NeedTripPostData } from "../../../services/ticket.type"
 import { TagConstIds } from "../../../services/types.tag"
 import { FormField } from "../FormField"
 import { ErrorMessage } from "@hookform/error-message"
+import { FormFeedback } from "./Feedback"
+import { useAddTransportTicket } from "./hooks"
 
 export type TransportNeededVariant = "whereFrom" | "whereTo"
 export type InputValuesType = {
@@ -38,6 +40,20 @@ export const FormNeedTransport = () => {
   const { data: locationTags = [] } = useQuery(`location-tags`, () => {
     return ticketService.locationTags()
   })
+  const addTicketMutation = useAddTransportTicket({
+    onSuccess: (rawResponse) => {
+      const { data } = rawResponse.data
+      const id = data.id
+
+      toast.success(translations["pages"]["add-ticket"]["need-added"])
+
+      setTimeout(() => {
+        return router.push(
+          RouteDefinitions.TicketDetails.replace(":id", String(id))
+        )
+      }, 1000)
+    },
+  })
 
   const mappedLocationTags = useMemo(() => {
     return locationTags
@@ -55,80 +71,7 @@ export const FormNeedTransport = () => {
       })
       .sort((a, b) => a.label.localeCompare(b.label))
   }, [locationTags])
-
-  const onSuccess = (rawResponse) => {
-    const { data } = rawResponse.data
-    const id = data.id
-
-    toast.success(translations["pages"]["add-ticket"]["need-added"])
-
-    setTimeout(() => {
-      return router.push(
-        RouteDefinitions.TicketDetails.replace(":id", String(id))
-      )
-    }, 1000)
-  }
-
-  const addTicketMutation = useMutation<
-    NeedTripPostData,
-    Error,
-    NeedTripPostData
-  >(
-    (newTicket) => {
-      const {
-        phone,
-        who,
-        what,
-        description,
-        adults,
-        children,
-        has_pets,
-        trip_when_text,
-        trip_when_date, // TODO:
-        trip_extra_luggage,
-        where_from_tag,
-        where_to_tag,
-      } = newTicket
-      const expirationTimestampSane = dayjs().add(24, "hour").format()
-
-      let when_date: string | undefined = undefined
-      if (trip_when_date != null) {
-        const date = new Date(trip_when_date)
-        when_date = date.toISOString()
-      }
-
-      const newTicketData = {
-        what,
-        description,
-        expirationTimestampSane,
-        phone,
-        who,
-        count: 0,
-        adults: adults ? adults : 0,
-        children: children ? children : 0,
-        has_pets: !has_pets ? "0" : "1",
-
-        // This is trip, so hardcore a trip tag
-        need_tag_id: [{ need_tag_id: { id: TagConstIds.transport } }],
-
-        // tripe specific
-        need_type: "trip",
-        where_to_tag,
-        where_from_tag,
-        trip_when_text,
-        trip_when_date: when_date,
-        trip_extra_luggage,
-      }
-
-      return axios.post(`/api/add-ticket`, newTicketData)
-    },
-    {
-      onSuccess,
-    }
-  )
-
   const useFormOptions = {}
-
   const {
     register,
     handleSubmit,
@@ -332,36 +275,7 @@ export const FormNeedTransport = () => {
             />
           </FormField>
 
-          <div>
-            {addTicketMutation.isError ? (
-              <div className="flex space-x-1 mt-2 items-center justify-center">
-                <div
-                  className="inline-flex items-center w-full place-content-center
-                py-1 border border-transparent shadow-sm text-sm
-                font-medium text-white bg-red-600
-                focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500
-                text-white cursor-default"
-                >
-                  {translations["errors"]["error-occured-while-adding"]}
-                  {addTicketMutation.error.message}
-                </div>
-              </div>
-            ) : null}
-
-            {addTicketMutation.isSuccess ? (
-              <div className="flex space-x-1 mt-2 items-center justify-center">
-                <div
-                  className="inline-flex items-center w-full place-content-center
-                py-1 border border-transparent shadow-sm text-sm
-                font-medium text-white bg-green-600
-                focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500
-                text-white cursor-default"
-                >
-                  {translations["pages"]["add-ticket"]["request-added"]}
-                </div>
-              </div>
-            ) : null}
-          </div>
+          <FormFeedback mutation={addTicketMutation} />
           <button
             type="submit"
             disabled={isDisabled}
